requirements.txt
# 核心依赖
SpeechRecognition==3.10.0
paho-mqtt==1.6.1
pyaudio==0.2.11
setuptools>=65.0.0
requests>=2.25.0

# 音频处理
librosa>=0.9.0
soundfile>=0.12.1
numpy>=1.21.0

# 数据处理
pandas>=1.3.0
matplotlib>=3.5.0

# 系统工具
psutil>=5.8.0
python-dotenv>=0.19.0

# 日志记录
loguru>=0.6.0

# 配置文件处理
PyYAML>=6.0
configparser>=5.2.0

# 时间处理
pytz>=2021.3

# AI语音识别可选依赖
# 百度AI语音识别（免费）
baidu-aip>=4.16.0

# Whisper本地语音识别（可选）
openai-whisper>=20230918
torch>=1.13.0
torchaudio>=0.13.0

# Azure语音服务（可选）
azure-cognitiveservices-speech>=1.24.0

# 谷歌语音识别（可选）
google-cloud-speech>=2.16.0

# 开发工具（开发环境）
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# 深度学习
torch
torchaudio
librosa